+++ SugarRecord/Source/CoreData/Storages/CoreDataiCloudStorage.swift
@@ -1,16 +1,7 @@
 import Foundation
 import CoreData
 
-public class CoreDataiCloudStorage: Storage {
-    
-    // MARK: - Attributes
-    
-    internal let store: CoreDataStore
-    internal var objectModel: NSManagedObjectModel! = nil
-    internal var persistentStore: NSPersistentStore! = nil
-    internal var persistentStoreCoordinator: NSPersistentStoreCoordinator! = nil
-    internal var rootSavingContext: NSManagedObjectContext! = nil
-
+public class CoreDataiCloudStorage: CoreDataBaseStorage, Storage {
     
     // MARK: - Storage
     
@@ -19,11 +10,9 @@ public class CoreDataiCloudStorage: Storage {
             return "CoreDataiCloudStorage"
         }
     }
-    public var type: StorageType = .coreData
+
     
-    public var mainContext: Context!
-    
-    public var saveContext: Context! {
+    override public var saveContext: Context! {
         get {
             let context = cdContext(withParent: .context(self.rootSavingContext), concurrencyType: .privateQueueConcurrencyType, inMemory: false)
             context.observe(inMainThread: true) { [weak self] (notification) -> Void in
@@ -33,78 +22,11 @@ public class CoreDataiCloudStorage: Storage {
         }
     }
     
-    public var memoryContext: Context! {
-        get {
-            let context =  cdContext(withParent: .context(self.rootSavingContext), concurrencyType: .privateQueueConcurrencyType, inMemory: true)
-            return context
-        }
-    }
-    
-    public func operation<T>(_ operation: @escaping (_ context: Context, _ save: @escaping () -> Void) throws -> T) throws -> T {
-        let context: NSManagedObjectContext = (self.saveContext as? NSManagedObjectContext)!
-        var _error: Error!
-        
-        var returnedObject: T!
-        
-        context.performAndWait {
-            do {
-                returnedObject = try operation(context, { () -> Void  in
-                    do {
-                        try context.save()
-                    }
-                    catch {
-                        _error = error
-                    }
-                    if self.rootSavingContext.hasChanges {
-                        self.rootSavingContext.performAndWait {
-                            do {
-                                try self.rootSavingContext.save()
-                            }
-                            catch {
-                                _error = error
-                            }
-                        }
-                    }
-                })
-            }
-            catch {
-                _error = error
-            }
-        }
-        if let error = _error {
-            throw error
-        }
-        
-        return returnedObject
-    }
-    
-    public func backgroundOperation(_ operation: @escaping (_ context: Context, _ save: @escaping () -> Void) -> (), completion: @escaping (Error?) -> ()) {
-        let context: NSManagedObjectContext = self.saveContext as! NSManagedObjectContext
-        var _error: Error!
-        context.perform {
-            operation(context, { () -> Void in
-                do {
-                    try context.save()
-                }
-                catch {
-                    _error = error
-                }
-                self.rootSavingContext.perform {
-                    if self.rootSavingContext.hasChanges {
-                        do {
-                            try self.rootSavingContext.save()
-                        }
-                        catch {
-                            _error = error
-                        }
-                    }
-                    completion(_error)
-                }
-            })
-        }
-    }
-    
+
     public func removeStore() throws {
+        guard let store = store else {
+            return
+        }
         try FileManager.default.removeItem(at: store.path() as URL)
     }
     
@@ -115,30 +37,15 @@ public class CoreDataiCloudStorage: Storage {
         try self.init(model: model, iCloud: iCloud, versionController: VersionController())
     }
     
-    internal init(model: CoreDataObjectModel, iCloud: CoreDataiCloudConfig, versionController: VersionController) throws {
-        self.objectModel = model.model()!
+    internal init(model: CoreDataObjectModel, iCloud: CoreDataiCloudConfig, versionController: VersionController) throws {        
+        super.init(model: model)
         self.persistentStoreCoordinator = NSPersistentStoreCoordinator(managedObjectModel: objectModel)
+        
         let result = try! cdiCloudInitializeStore(storeCoordinator: persistentStoreCoordinator, iCloud: iCloud)
         self.store = result.0
         self.persistentStore = result.1
-        self.rootSavingContext = cdContext(withParent: .coordinator(self.persistentStoreCoordinator), concurrencyType: .privateQueueConcurrencyType, inMemory: false)
-        self.mainContext = cdContext(withParent: .context(self.rootSavingContext), concurrencyType: .mainQueueConcurrencyType, inMemory: false)
         self.observeiCloudChangesInCoordinator()
-        #if DEBUG
-        versionController.check()
-        #endif
     }
-    
-    
-    // MARK: - Public
-
-#if os(iOS) || os(tvOS) || os(watchOS)
-    
-    public func observable<T: NSManagedObject>(request: FetchRequest<T>) -> RequestObservable<T> where T:Equatable {
-        return CoreDataObservable(request: request, context: self.mainContext as! NSManagedObjectContext)
-    }
-    
-#endif
     
     // MARK: - Private
